+++ adapter-module/src/main/kotlin/com/reservation/rest/user/general/attribute/change/password/GeneralUserChangePasswordController.kt
@@ -1,4 +1,4 @@
-package com.reservation.rest.user.general.attribute
+package com.reservation.rest.user.general.attribute.change.password
 
 import com.reservation.rest.common.response.BooleanResponse
 import com.reservation.rest.user.general.GeneralUserUrl
