+++ adapter-module/src/test/kotlin/com/reservation/user/general/attribute/GeneralUserChangeNicknameControllerTest.kt
@@ -7,7 +7,7 @@ import com.reservation.config.restdoc.PathParameter
 import com.reservation.config.restdoc.RestDocuments
 import com.reservation.config.security.TestSecurity
 import com.reservation.fixture.CommonlyUsedArbitraries
-import com.reservation.rest.user.general.attribute.GeneralUserChangeNicknameController
+import com.reservation.rest.user.general.attribute.change.nickname.GeneralUserChangeNicknameController
 import com.reservation.rest.user.general.request.GeneralUserChangeNicknameRequest
 import com.reservation.user.self.port.input.ChangeGeneralUserNicknameUseCase
 import com.reservation.utilities.generator.uuid.UuidGenerator
